[GPU] [OverClock] Add extra freq 750mhz:
Stock is 710mhz:
We added another freq, and made it maxed out at 750mhz,
This is no way the final max, but its a good start for adding
another freq to GPU contoler.

After adding this, Make sure you add the commit that
will limit the GPU to stock defaul freq of 710mhz
durning boot. That will insure stability of booting..

Signed-off-by: Eliminater74 <eliminater74@gmail.com>